change balance from int to float64

+++ pkg/users/user.go
@@ -2,8 +2,8 @@ package users
 
 // User represents a user
 type User struct {
-	ID        string `json:"ID"`
-	Balance   int    `json:"balance"`
-	Currency  string `json:"currency"`
-	CreatedAt string `json:"createdat"`
+	ID        string  `json:"ID"`
+	Balance   float64 `json:"balance"`
+	Currency  string  `json:"currency"`
+	CreatedAt string  `json:"createdat"`
 }
